# src/main/java/com/example/bookstore/model/BookModel.java
package com.example.bookstore.model;

import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.Id;
import jakarta.persistence.Table;

import java.math.BigDecimal;
import java.util.UUID;

@Entity
@Table
public class BookModel {
    @Id
    @GeneratedValue
    private UUID id;
    private String name;
    private String code;
    private String publisher;
    private String author;
    private BigDecimal quantity;

    public BookModel() {
    }

    public UUID getId() {
        return id;
    }

    public void setId(UUID id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getCode() {
        return code;
    }

    public void setCode(String code) {
        this.code = code;
    }

    public String getPublisher() {
        return publisher;
    }

    public void setPublisher(String publisher) {
        this.publisher = publisher;
    }

    public String getAuthor() {
        return author;
    }

    public void setAuthor(String author) {
        this.author = author;
    }

    public BigDecimal getQuantity() {
        return quantity;
    }

    public void setQuantity(BigDecimal initialQuantity) {
        this.quantity = initialQuantity;
    }

    public BookModel(String name, String code, String publisher, String author, BigDecimal initialQuantity) {
        this.name = name;
        this.code = code;
        this.publisher = publisher;
        this.author = author;
        this.quantity = initialQuantity;
    }

    @Override
    public String toString() {
        return "BookModel{" +
                "id='" + id + '\'' +
                ", name='" + name + '\'' +
                ", code='" + code + '\'' +
                ", publisher='" + publisher + '\'' +
                ", author='" + author + '\'' +
                ", quantity=" + quantity +
                '}';
    }
}